Changed all gaussian count related types to unsigned char

+++ GaussMixDetector.cpp
@@ -82,7 +82,7 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 		const auto* ptF = fClone.ptr<double>(i);
 		auto* ptMo = motion.ptr<uchar>(i);
 		auto* ptK = currentK.ptr<uchar>(i);
-		for ( int_fast16_t k = 0; k < K; k++ )
+		for ( uchar k = 0U; k < K; k++ )
 		{
 			ptM[k] = mean[k].ptr<double>(i);
 			ptD[k] = deviation[k].ptr<double>(i);
@@ -91,16 +91,16 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 		for ( int j = 0; j < fCols*fChannels; j += fChannels )
 		{
 			double tmpF = ptF[j];
-			int_fast16_t tmpK = ptK[j];
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			uchar tmpK = ptK[j];
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				tmpM.at(k) = ptM[k][j];
 				tmpD.at(k) = ptD[k][j];
 				tmpW.at(k) = ptW[k][j];
 			}
 
-			int_fast16_t count = 0;
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			uchar count = 0U;
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				isCurrent[k] = false;
 				if (( tmpF < tmpM.at(k) + tmpD.at(k) ) && ( tmpF > tmpM.at(k) - tmpD.at(k) ))
@@ -113,7 +113,7 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 					isCurrent[k] = false;
 				}
 			}
-			if( count == 0 )
+			if( count == 0U )
 			{
 				if ( tmpK < K )
 				{
@@ -133,7 +133,7 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 			}
 			else
 			{
-				for ( int_fast16_t k = 0; k < tmpK; k++ )
+				for ( uchar k = 0U; k < tmpK; k++ )
 				{
 					if ( isCurrent[k] ) {
 						double r = alpha * normDistrib( tmpF, tmpM.at(k), tmpD.at(k) );
@@ -150,19 +150,19 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 			}
 
 			double w = 0;
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				tmpW.at(k) = tmpW.at(k) * (1-alpha) + alpha*int(isCurrent[k]);
 				w += tmpW.at(k);
 			}
 
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				tmpW.at(k) = tmpW.at(k) / w;
 			}
 
 			w = 0;
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				w += tmpW.at(k);
 			}
@@ -173,13 +173,13 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 			while (!noMov)
 			{
 				noMov = true;
-				for ( int_fast16_t k = 0; k < tmpK-1; k++ )
+				for ( uchar k = 0U; k < tmpK-1U; k++ )
 				{
-					if ( tmpW.at(k) > tmpW.at(k+1) )
+					if ( tmpW.at(k) > tmpW.at(k+1U) )
 					{
-						std::swap(tmpW.at(k), tmpW.at(k+1));
-						std::swap(tmpM.at(k), tmpM.at(k+1));
-						std::swap(tmpD.at(k), tmpD.at(k+1));
+						std::swap(tmpW.at(k), tmpW.at(k+1U));
+						std::swap(tmpM.at(k), tmpM.at(k+1U));
+						std::swap(tmpD.at(k), tmpD.at(k+1U));
 
 						noMov = false;
 					}
@@ -188,10 +188,10 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 
 			int motionI = j / fChannels;
 
-			int_fast16_t count2 = 0;
+			uchar count2 = 0U;
 			w = tmpW.at(0);
 
-			for ( count2 = 1; count2 < tmpK; count2++ )
+			for ( count2 = 1U; count2 < tmpK; count2++ )
 			{
 				if ( w >= Cf ) {
 					break;
@@ -200,19 +200,19 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 				w += tmpW.at(count2);
 			}
 
-			for ( int_fast16_t k = 0; k < count2; k++ )
+			for ( uchar k = 0U; k < count2; k++ )
 			{
 				if(isCurrent[k])
 				{
-					ptMo[motionI] = 255;
+					ptMo[motionI] = 255U;
 				}
 			}
 
 
-			if( tmpK > 1 )
+			if( tmpK > 1U )
 			{
 				bool isMotion = true;
-				for ( int_fast16_t k = 1; k < tmpK; k++ )
+				for ( uchar k = 1U; k < tmpK; k++ )
 				{
 					if( !(abs(tmpM.at(0) - tmpM.at(k)) > T) )
 					{
@@ -223,12 +223,12 @@ void GaussMixDetector::getpwUpdateAndMotion( cv::Mat& motion )
 
 				if( isMotion )
 				{
-					ptMo[motionI] = 255;
+					ptMo[motionI] = 255U;
 				}
 			}
 
-			ptK[j] = static_cast<uchar>(tmpK);
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			ptK[j] = tmpK;
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				ptM[k][j] = tmpM.at(k);
 				ptD[k][j] = tmpD.at(k);
@@ -270,14 +270,14 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 		const auto* ptF = fClone.ptr<double>(i);
 		auto* ptMo = motion.ptr<uchar>(i);
 		auto* ptK = currentK.ptr<uchar>(i);
-		for ( int_fast16_t k = 0; k < K; k++ )
+		for ( uchar k = 0U; k < K; k++ )
 		{
 			ptM.at(k) = mean[k].ptr<double>(i);
 			ptD.at(k) = deviation[k].ptr<double>(i);
 			ptW.at(k) = weight[k].ptr<double>(i);
 		}
 
-		int_fast16_t tmpK = 0;
+		uchar tmpK = 0U;
 		for ( int j = 0; j < fCols; j++ )
 		{
 			const int iRGB = j*fChannels;
@@ -289,7 +289,7 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 			}
 			tmpK = ptK[j];
 
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				for ( int c = 0; c < 3; c++ )
 				{
@@ -303,8 +303,8 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 				delta.at(k) = tmpF - tmpM.at(k);
 			}
 
-			int_fast16_t count = 0;
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			uchar count = 0U;
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				isCurrent.at(k) = false;
 				if ( Mahalanobis(delta.at(k), tmpD.at(k)) < sqrt( cv::trace(tmpD.at(k)) ) )
@@ -318,7 +318,7 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 				}
 			}
 
-			if( count == 0 )
+			if( count == 0U )
 			{
 				if ( tmpK < K )
 				{
@@ -329,14 +329,14 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 				}
 				else
 				{
-					tmpM.at(K-1) = tmpF;
-					tmpD.at(K-1) = initDeviation*cv::Matx33d::eye();
-					tmpW.at(K-1) = alpha;
+					tmpM.at(K-1U) = tmpF;
+					tmpD.at(K-1U) = initDeviation*cv::Matx33d::eye();
+					tmpW.at(K-1U) = alpha;
 				}
 			}
 			else
 			{
-				for ( int_fast16_t k = 0; k < tmpK; k++ )
+				for ( uchar k = 0U; k < tmpK; k++ )
 				{
 					if ( isCurrent.at(k) )
 					{
@@ -349,13 +349,13 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 
 			{
 				double w = 0;
-				for ( int_fast16_t k = 0; k < tmpK; k++ )
+				for ( uchar k = 0U; k < tmpK; k++ )
 				{
 					tmpW.at(k) = tmpW.at(k) * (1-alpha) + alpha*int(isCurrent.at(k));
 					w += tmpW.at(k);
 				}
 
-				for ( int_fast16_t k = 0; k < tmpK; k++ )
+				for ( uchar k = 0U; k < tmpK; k++ )
 				{
 					tmpW.at(k) = tmpW.at(k) / w;
 				}
@@ -365,13 +365,13 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 			while (!noMov)
 			{
 				noMov = true;
-				for ( int_fast16_t k = 0; k < tmpK-1; k++ )
+				for ( uchar k = 0U; k < tmpK-1U; k++ )
 				{
-					if ( tmpW.at(k) < tmpW.at(k+1) )
+					if ( tmpW.at(k) < tmpW.at(k+1U) )
 					{
-						std::swap(tmpW.at(k), tmpW.at(k+1));
-						cv::swap(tmpM.at(k), tmpM.at(k+1));
-						cv::swap(tmpD.at(k), tmpD.at(k+1));
+						std::swap(tmpW.at(k), tmpW.at(k+1U));
+						cv::swap(tmpM.at(k), tmpM.at(k+1U));
+						cv::swap(tmpD.at(k), tmpD.at(k+1U));
 
 						noMov = false;
 					}
@@ -382,11 +382,11 @@ void GaussMixDetector::getpwUpdateAndMotionRGB( cv::Mat& motion )
 
 			if( FG )
 			{
-				ptMo[j] = 255;
+				ptMo[j] = 255U;
 			}
 
-			ptK[j] = static_cast<uchar>(tmpK);
-			for ( int_fast16_t k = 0; k < tmpK; k++ )
+			ptK[j] = tmpK;
+			for ( uchar k = 0U; k < tmpK; k++ )
 			{
 				for ( int c = 0; c < fChannels; c++ )
 				{
